tighter handling of object aliases
previously, partial objects in the store would be copied for
on-stack symbols. this lead to putObject(p_X, ..) modifying only
one alias of a shared object, causing lots of failed null
checks in test C, because writes were not visible and
fields looked like they were null.

now we're getting an ArrayIndexOutOfBoundsException instead of
NPEs, whoopey!